Added: A bunch of functions to help with debugging systems and behavior trees, like recording game states, reseting to previous game states, etc.
Fixed:
Entities would get in a really weird state of data when their behavior
tree was checking to see if it was running the highest priority node.
Because one of the nodes had a randomness to its output, when the tree
is run again to determine the highest priority node, there was a chance
that the "PickRandomLocation" node would pick an unpassable tile,
causing it to fail. I never anticipated something failing when testing
the highest priority node.
Anyways, now failures are just ignored and the tree basically doesn't
get to do a highest priority check for that frame.